CM-395 - Updated Subfloor to set working directory for JUnit
When being called from one of the multi-build files, JUnit
sets the working directory of the sub-ant files to the directory
of the outermost file.

The proper fix for that is to change the shared "subfloor.xml"
file and to change the 'junit' tasks within. They need a
attribute 'dir' with the value of '${basedir}' to set the
working directory of the tests to the module directory.

If the working directory is not right, many unit tests will fail
as they cannot find their SQL datasource and/or gold sample
directory.

I changed all reporting 'subfloor.xml' files to contain that
property. However, this must be part of the global subfloor
file or our local changes get wiped out when the master subfloor
file is pushed to all known locations.