PRD-4653: Fix and test-case.
We needed to filter out state-markers from inline-subreports. Inline subreports are processed in bulk, without the report-processor seeing the intermediate states. Therefore the report processor cannot revert back to any of these states, only to the state that triggered the subreport processing. By filtering, we now return that state instead and therefore the pagebreak code can jump in at the right time.