PRD-4769 - Prepare crosstab editing
Cleanup the API so that that report-render-context is hidden.
This separates the public usage of the context data from the
internal processing and will make implementing undo and redo
a lot easier.

Separated report-designer container from report-context, so
that the outside container knows less about the internal workings
of the report-context.