[PDI-12579] - PDI Report Wizard doesn't display the field names properly
The cause of the issue is that SwtList component does not properly sets Label for items (updateLabel() method)
However taking into account the risk of modifying the core Xul functionality it's been decided to
did the changes in SourceFieldDefinition class.

The fix is to remove an empty implementation of XulEvenSource making Xul process the added ListItems differently.

Tested manually. There are no unit tests since this is pure UI change.